move site in core.local
