+++ gl2f/local/__init__.py
@@ -23,111 +23,6 @@ def clear_cache():
 		for i in os.listdir(d):
 			os.remove(os.path.join(d, i))
 
-def declare_js_string_constant(name, value):
-	return f'const {name} = "{value}";\n'
-
-def install_to(dst, link_type):
-	import shutil, socket
-
-	src = local.fs.package_data('site')
-
-	if os.path.exists(dst):
-		print(f'reinstalling {dst} that already exists')
-		rm = shutil.rmtree if os.path.isdir(dst) else os.remove
-		rm(dst)
-
-	cp = shutil.copytree if os.path.isdir(src) else shutil.copyfile
-	cp(src, dst)
-
-	if link_type == 'symbolic':
-		contents_path = 'contents'
-		index_path = 'index.js'
-		os.symlink(local.fs.refdir('contents'), os.path.join(dst, 'contents'))
-		os.symlink(os.path.join(local.fs.home(), 'index.js'), os.path.join(dst, 'index.js'))
-	elif link_type == 'relative':
-		contents_path = '../contents'
-		index_path = '../index.js'
-
-	else:
-		raise RuntimeError('unknown path type')
-
-	index.main(site=dst, full=True)
-
-	with open(os.path.join(dst, 'constants.js'), 'w', encoding='utf-8') as f:
-		f.write(declare_js_string_constant('hostname', config['host-name']))
-		f.write(declare_js_string_constant('contentsPath', contents_path))
-		f.write(declare_js_string_constant('indexPath', index_path))
-
-
-	print(f'installed site into {dst}')
-
-
-class index:
-	@staticmethod
-	def load():
-		try:
-			path = os.path.join(local.fs.refdir_untouch('site'), 'index.js')
-			with open(path, encoding='utf-8') as f:
-				raw = f.read()
-			return json.loads(re.sub(r'^.+?=', '', raw))
-
-		except:
-			return {}
-
-
-	@staticmethod
-	def value(i):
-		from ..core import board, article
-
-		item = local.data.load(i)
-		media = [i for i, _ in article.ptn_media.findall(item['values']['body'])]
-		return {
-			'title': item['values']['title'],
-			'board': board.get('id', item['boardId'])['page'],
-			'author': item.get('category', {'name':''})['name'],
-			'date': item['openingAt'],
-			'media': [''.join(x) for x in sorted(
-				filter(lambda x:x[0] in media,
-					(os.path.splitext(i) for i in local.fs.listdir(os.path.join('contents', i)))
-				),
-				key=lambda x:media.index(x[0])
-			)],
-			'expired': item.get('closingAt', None),
-			'body': build_body(item)
-		}
-
-
-	@staticmethod
-	def create_table(contents):
-		from concurrent.futures import ThreadPoolExecutor
-		with ThreadPoolExecutor() as e:
-			values = e.map(index.value, contents)
-		return {k:v for k, v in zip(contents, values)}
-
-
-	@staticmethod
-	def main(site=None, full=False):
-		if not site:
-			site = local.fs.refdir_untouch('site')
-
-		if not site:
-			print('site not installed. return')
-			return
-
-		if full:
-			table = index.create_table(local.fs.listdir('contents'))
-		else:
-			prev = index.load()
-			contents = list(set(local.fs.listdir('contents')).difference(prev.keys()))
-			table = prev | index.create_table(contents)
-
-		out = os.path.join(local.fs.home(), 'index.js')
-		with open(out, 'w', encoding='utf-8') as f:
-			print(f'const table={json.dumps(table, separators=(",", ":"), ensure_ascii=False)}', file=f)
-
-		print(f'saved {out}')
-
-
 def open_site():
 	import webbrowser
 
@@ -135,82 +30,13 @@ def open_site():
 
 	if not os.path.exists(html):
 		if 'n' != input('site not installed. install now? (Y/n)').lower():
-			install_to(os.path.join(local.fs.home(), 'site'), 'relative')
+			local.site.install_to(os.path.join(local.fs.home(), 'site'), 'relative')
 		else:
 			return
 	else:
-		index.main()
+		local.site.index.main()
 
 	webbrowser.open(f'file://{html}')
-
-def build_body(item):
-	from ..core import article
-
-	i = item['contentId']
-	media_list = local.fs.listdir(f'contents/{i}')
-
-	def up(match):
-		m, t = match.groups()
-		try:
-			p = next(p for p in media_list if p.startswith(m))
-		except:
-			return ''
-
-		if t == 'image':
-			return f'<img src=../contents/{i}/{p}></img>'
-		elif t == 'video':
-			return f'<video controls autoplay muted loop src=../contents/{i}/{p}></video>'
-		else:
-			return ''
-
-	return article.ptn_media.sub(up, item['values']['body'])
-
-
-def get_local_ip():
-	import socket
-	s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-	try:
-		s.connect(('8.8.8.8', 80))
-		return s.getsockname()[0]
-	except:
-		return 'localhost'
-	finally:
-		s.close()
-
-def serve(port, browse=False):
-	import http.server, socketserver, socket, threading
-	import webbrowser
-	import tempfile
-
-	webbrowser.register("termux-open '%s'", None)
-
-	with tempfile.TemporaryDirectory() as tmp:
-		site = os.path.join(tmp, 'site')
-		print(site)
-		install_to(site, 'symbolic')
-
-		class Handler(http.server.SimpleHTTPRequestHandler):
-			def __init__(self, *args, **kwargs):
-				super().__init__(*args, directory=site, **kwargs)
-
-			def end_headers(self):
-				self.send_header('Accept-Ranges', 'bytes')
-				super().end_headers()
-
-		url = f'http://{get_local_ip()}:{port}'
-		with socketserver.ThreadingTCPServer(('0.0.0.0', port), Handler) as httpd:
-			print(f'serving at {url}')
-
-			server_thread = threading.Thread(target=httpd.serve_forever)
-			server_thread.daemon = True
-			server_thread.start()
-
-			try:
-				if browse:
-					webbrowser.open(url)
-				server_thread.join()
-			except  KeyboardInterrupt:
-				pass
 
 
 def add_to():
@@ -233,8 +59,8 @@ def add_args(parser):
 	p.add_argument('archive')
 	p.set_defaults(handler=lambda args:archive.import_contents(args.archive))
 
-	sub.add_parser('index', description='Create index of contents for web viewer').set_defaults(handler = lambda _:index.main(full=True))
-	sub.add_parser('install', description='Install static web viewer').set_defaults(handler=lambda _:install_to(os.path.join(local.fs.home(), 'site'), 'relative'))
+	sub.add_parser('index', description='Create index of contents for web viewer').set_defaults(handler = lambda _:local.site.index.main(full=True))
+	sub.add_parser('install', description='Install static web viewer').set_defaults(handler=lambda _:local.site.install_to(os.path.join(local.fs.home(), 'site'), 'relative'))
 
 	p = sub.add_parser('ls', description='List all local contents')
 	p.add_argument('--order', type=str,
@@ -251,7 +77,7 @@ def add_args(parser):
 		help='Set port to host on')
 	p.add_argument('--open', action='store_true',
 		help='Also open in the browser')
-	p.set_defaults(handler=lambda args:serve(args.port, args.open))
+	p.set_defaults(handler=lambda args:local.site.serve(args.port, args.open))
 
 	return sub
